Refactor KEM speed test
Rename and change package for readability. Some problems in the pom.xml still remains but it is easy to execute (e.g., mvn -exec, IDE, etc.) but maybe a better packaging should be used. Other things can be renamed (such as "Classical") but for now I believe the next steps are adding signatures to this benchmarks.
