+++ pom.xml
@@ -9,12 +9,8 @@
         <dependency>
             <groupId>org.bouncycastle</groupId>
             <artifactId>bcprov-jdk18on</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.bouncycastle</groupId>
-            <artifactId>bc-fips</artifactId>
-            <version>1.0.2.4</version>
-            <type>jar</type>
+            <version>1.77</version>
+                <type>jar</type>
         </dependency>
         <dependency>
             <groupId>info.picocli</groupId>
@@ -22,21 +18,62 @@
             <version>4.1.4</version>
             <type>jar</type>
         </dependency>
+        <!--<dependency>
+            <groupId>org.bouncycastle</groupId>
+            <artifactId>bc-fips</artifactId>
+            <version>1.0.2.4</version>
+            <type>jar</type>
+        </dependency>
+        -->
     </dependencies>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
-        <exec.mainClass>grupopqc.utfprtd.speedtesting.KEMSpeedTest</exec.mainClass>
+        <exec.mainClass>grupopqc.utfprtd.hybridexample.KEMSpeedTest</exec.mainClass>
     </properties>
-    <dependencyManagement>
+   
+   <!--<build>
+        <plugins>
+       <plugin>
+           <groupId>org.apache.maven.plugins</groupId>
+           <artifactId>maven-shade-plugin</artifactId>
+           <version>3.0.0</version>
+           <executions>
+               <execution>
+                   <phase>package</phase>
+                   <goals>
+                       <goal>shade</goal>
+                   </goals>
+                   <configuration>
+                       <transformers>
+                           <transformer
+                               implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                               <manifestEntries>
+                                   <Main-Class>grupopqc.utfprtd.hybridexample.KEMSpeedTest</Main-Class>
+                               </manifestEntries>
+                           </transformer>
+                       </transformers>
+                   </configuration>
+               </execution>
+           </executions>
+       </plugin>
+       </plugins>
+    </build> -->
+    <!--<dependencyManagement>
         <dependencies>            
             <dependency>
                 <groupId>org.bouncycastle</groupId>
                 <artifactId>bcprov-jdk18on</artifactId>
                 <version>1.77</version>
                 <type>jar</type>
-            </dependency>            
+            </dependency>           
+            <dependency>
+                <groupId>info.picocli</groupId>
+                <artifactId>picocli</artifactId>
+                <version>4.1.4</version>
+                <type>jar</type>
+            </dependency>
         </dependencies>
-    </dependencyManagement>        
+    </dependencyManagement>        -->
 </project>
